requirements.txt
gradio>=4.0.0
google-generativeai>=0.3.0
fpdf>=1.7.2